% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_index.R
\docType{package}
\name{MassSpecUtils}
\alias{MassSpecUtils-package}
\alias{MassSpecUtils}
\title{MassSpecUtils}
\description{
提供一系列用于处理 DIA-NN 输出数据的工具函数，
包括肽段位置查找、修饰位点解析等功能。
}
\seealso{
Useful links:
\itemize{
  \item \url{https://zhbhmile.github.io/MassSpecUtils}
  \item \url{https://github.com/ZHBHSMILE/MassSpecUtils}
  \item Report bugs at \url{https://github.com/ZHBHSMILE/MassSpecUtils/issues}
}

}
\author{
\strong{Maintainer}: Zhaopeng Lv \email{lvzhaopenglv@126.com}

}
\keyword{internal}
